+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/batch/SearchableJobInstanceDao.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.batch.admin.service;
+package org.springframework.cloud.dataflow.server.batch;
 
 import org.springframework.batch.core.repository.dao.JobInstanceDao;
