+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/batch/SimpleJobServiceFactoryBean.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.batch.admin.service;
+package org.springframework.cloud.dataflow.server.batch;
 
 import static org.springframework.batch.support.DatabaseType.SYBASE;
